orchestrator/requirements.txt
aio-pika==9.5.4
anthropic==0.42.0
kubernetes==31.0.0
jinja2==3.1.5
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
python-json-logger==2.0.7
